requirements.txt
# requirements.txt
langgraph==0.0.44
langchain==0.1.0
langchain-mistralai==0.1.0
mistralai>=0.1.3
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
jinja2>=3.1.3
python-json-logger>=2.0.7
termcolor>=2.3.0
tqdm>=4.66.0
fastapi>=0.104.1
uvicorn>=0.24.0
numpy>=1.24.0